% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{crossword}
\alias{crossword}
\title{Title}
\description{
Description
}
\section{Usage}{

\preformatted{
  cw <- crossword$new()
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{process} object.}
}
}

\section{Details}{

\code{$new()} starts a new process, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.

\code{$is_alive()} checks if the process is alive. Returns a logical
scalar.
}

\examples{
p <- crossword$new(rows = 10, columns = 10)

}
