% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_glue_clues.R
\name{cw_glue_clues}
\alias{cw_glue_clues}
\title{glue together parameters and cwpuzzle template}
\usage{
cw_glue_clues(cw)
}
\arguments{
\item{cw}{}
}
\description{
glue together parameters and cwpuzzle template
}
