% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_glue_letters.R
\name{cw_glue_letters}
\alias{cw_glue_letters}
\title{convert to cwpuzzle table}
\usage{
cw_glue_letters(x)
}
\arguments{
\item{x}{the letters matrix of a crossword}
}
\description{
Converts an R-crossword to a table that can be used with LaTeX cwpuzzle
}
