% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_glue_solution.R
\name{cw_glue_solution}
\alias{cw_glue_solution}
\title{glue together parameters and cwpuzzle template}
\usage{
cw_glue_solution(cw)
}
\arguments{
\item{cw}{}
}
\description{
glue together parameters and cwpuzzle template
}
