% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_glue_table.R
\name{cw_glue_table}
\alias{cw_glue_table}
\title{convert to cwpuzzle}
\usage{
cw_glue_table(x)
}
\arguments{
\item{x}{the letters matrix of a crossword}
}
\description{
Converts an R-crossword to a table that can be used with LaTeX cwpuzzle
}
