% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_normalize_words.R
\name{cw_normalize_words}
\alias{cw_normalize_words}
\title{normalize words to be added to grid}
\usage{
cw_normalize_words(words)
}
\arguments{
\item{words}{character vector of words to normalize for crossword usage}
}
\description{
normalize words to be added to grid
}
