% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_r6_extended.R
\docType{class}
\name{cw_r6_extended}
\alias{cw_r6_extended}
\title{extended R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
cw_r6_extended
}
\value{
Object of \code{\link{R6Class}}
}
\description{
extended R6 class
}
\keyword{data}
