% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_to_json.R
\name{cw_to_json}
\alias{cw_to_json}
\title{function implementing to_json method}
\usage{
cw_to_json(cw, pretty = FALSE)
}
\arguments{
\item{cw}{an object of class crossword}

\item{pretty}{should json formatted to be mor human readable or not}
}
\description{
function implementing to_json method
}
