% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_to_pdf.R
\name{cw_to_pdf}
\alias{cw_to_pdf}
\title{function that takes a crossword and renders it as PDF}
\usage{
cw_to_pdf(cw, fname = NULL)
}
\description{
function that takes a crossword and renders it as PDF
}
