% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_words_add_id.R
\name{cw_words_add_id}
\alias{cw_words_add_id}
\title{add an id unique for all unique combinations of row and column}
\usage{
cw_words_add_id(cw_words)
}
\arguments{
\item{cw_words}{}
}
\description{
add an id unique for all unique combinations of row and column
}
