% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{greplv}
\alias{greplv}
\title{a vectorized version of grep}
\usage{
greplv(pattern, x, ignore.case = FALSE, perl = FALSE, fixed = FALSE,
  useBytes = FALSE)
}
\description{
a vectorized version of grep
}
