% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{function that turn matrix into a data.frame in long format}
\usage{
matrix_to_df(x)
}
\arguments{
\item{x}{the data.frame to transform}
}
\description{
function that turn matrix into a data.frame in long format
}
