% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{normalize_words}
\alias{normalize_words}
\title{normalize words to be added to grid}
\usage{
normalize_words(words)
}
\arguments{
\item{words}{character vector of words to normalize for crossword usage}
}
\description{
normalize words to be added to grid
}
