% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6extended.R
\docType{class}
\name{r6extended}
\alias{r6extended}
\title{extended R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
r6extended
}
\value{
Object of \code{\link{R6Class}}
}
\description{
extended R6 class
}
\keyword{data}
