% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_wordlist_animal_en.R
\docType{data}
\name{cw_wordlist_animal_en}
\alias{cw_wordlist_animal_en}
\title{en - animals}
\format{An object of class \code{data.frame} with 68 rows and 2 columns.}
\usage{
cw_wordlist_animal_en
}
\description{
data frame of words and clues
}
\keyword{datasets}
