% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{crossword}
\alias{crossword}
\alias{crossword}
\title{Title}
\description{
Description

dings
}
\section{Usage}{

\preformatted{
  cw <- crossword$new()
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{process} object.}
}
}

\section{Details}{

\code{$new()} starts a new process, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.

\code{$is_alive()} checks if the process is alive. Returns a logical
scalar.
}

\examples{
cw <- crossword$new(rows = 4, columns = 4)
cw$add_words(c("back", "nasa", "kick", "nuk", "ic", "sic"))
cw
cw$letters
cw$words

}
