% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossword.R
\docType{data}
\name{Crossword}
\alias{Crossword}
\title{A class for generating crosswords from word lists}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Crossword
}
\description{
Description
}
\section{Usage}{

\preformatted{
  cw <- Crossword$new()
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{process} object.}
}
}

\section{Details}{

\code{$new()} starts a new process, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.

\code{$is_alive()} checks if the process is alive. Returns a logical
scalar.
}

\examples{

library(crossword)
cw <- Crossword$new(rows = 4, columns = 4)
cw$add_words(c("back", "nasa", "kick", "nuk", "ic", "sic"))
cw
cw$letters
cw$words

}
\keyword{datasets}
