% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_glue_head.R
\name{cw_glue_head}
\alias{cw_glue_head}
\title{glue together parameters and cwpuzzle template}
\usage{
cw_glue_head(cw, geometry_options = "a4paper, margin=2cm")
}
\arguments{
\item{cw}{}
}
\description{
glue together parameters and cwpuzzle template
}
